[refined] move stake-change-indicator to its own top-level component dir
