+++ src/components/stake-change-indicator/stake-change-indicator.tsx
